# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: BMKG Weather API
  description: |
    REST API untuk mengambil data cuaca real-time dari stasiun cuaca BMKG Indonesia.
    
    ## Tipe Stasiun yang Didukung
    - **AWS** (Automatic Weather Station) - Stasiun cuaca lengkap
    - **AAWS** (Advanced AWS) - AWS dengan sensor tambahan
    - **ARG** (Automatic Rain Gauge) - Pengukur curah hujan otomatis
    - **ASRS** (Automatic Solar Radiation Station) - Stasiun radiasi matahari
    - **Soil** - Stasiun kelembaban tanah
    - **Iklimmikro** - Stasiun iklim mikro multi-level (4m, 7m, 10m)
    
    ## Format Response
    - **JSON** (default) - Format standar dengan metadata
    - **GeoJSON** - Format RFC 7946 untuk mapping/visualisasi
    
    ## Client Library
    Gunakan `weather-client` package untuk konversi data:
    
    **Public Weather:**
    - `publicToGeoJSON()` - Convert ke GeoJSON
    - `filterPublicGeoJSON()` - Filter by province/kabupaten/kecamatan
    - `filterPublicByBoundingBox()` - Filter by bounding box
    
    **AWS Stations:**
    - `awsToGeoJSON()` - Convert ke GeoJSON
    - `awsToGeoJSONFeature()` - Convert single station
  version: 1.0.0
  contact:
    name: BMKG Tools
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /public/nowcasting:
    get:
      tags:
        - Public
      summary: Get Nowcasting Data
      description: |
        Mendapatkan data nowcasting (prakiraan cuaca jangka pendek).
        
        Sumber data:
        - `signature` (default) - dari signature.bmkg.go.id (JSON)
        - `xml` / `databmkg` - dari www.bmkg.go.id (XML → JSON), memerlukan parameter `province`
      operationId: getNowcasting
      parameters:
        - name: type
          in: query
          description: Tipe sumber data
          schema:
            type: string
            enum: [signature, xml, databmkg]
            default: signature
        - name: code
          in: query
          description: Kode stasiun (untuk type=signature)
          schema:
            type: string
            default: CJH
          example: CJH
        - name: province
          in: query
          description: |
            Nama provinsi (WAJIB untuk type=xml/databmkg).
            Gunakan underscore `_` untuk spasi.
          schema:
            type: string
          example: jawa_tengah
      responses:
        '200':
          description: Data nowcasting berhasil diambil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NowcastingResponse'
        '400':
          description: Parameter tidak valid (missing province, unsupported type)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Data tidak ditemukan untuk provinsi yang diminta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /public/location:
    get:
      tags:
        - Public
      summary: Get Weather by Location
      description: |
        Mendapatkan data cuaca untuk lokasi spesifik berdasarkan kode ADM4 atau koordinat.
        
        **Pilih salah satu:**
        - `code` - Kode wilayah ADM4 (contoh: 33.01.22.1003 untuk Cilacap Tengah)
        - `lat` + `lon` - Koordinat latitude dan longitude
      operationId: getLocationWeather
      parameters:
        - name: code
          in: query
          description: |
            Kode wilayah ADM4 (format: XX.XX.XX.XXXX)
            Contoh: `33.01.22.1003` untuk Cilacap Tengah, Sidanegara
          schema:
            type: string
          example: '33.01.22.1003'
        - name: lat
          in: query
          description: Latitude koordinat (wajib bersama `lon`)
          schema:
            type: number
            minimum: -90
            maximum: 90
          example: -7.656747
        - name: lon
          in: query
          description: Longitude koordinat (wajib bersama `lat`)
          schema:
            type: number
            minimum: -180
            maximum: 180
          example: 109.115523
      responses:
        '200':
          description: Data lokasi berhasil diambil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocationWeatherResponse'
        '400':
          description: Parameter tidak valid atau tidak lengkap
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /public/weather:
    get:
      tags:
        - Public
      summary: Get Public Weather Data (Bulk)
      description: |
        Mendapatkan data prakiraan cuaca publik untuk banyak lokasi sekaligus.
        Data dapat difilter berdasarkan provinsi, kabupaten, atau kecamatan.
        
        **Gunakan endpoint ini untuk:**
        - Menampilkan peta dengan banyak titik cuaca
        - Mengambil data seluruh wilayah
        
        **Untuk data lokasi spesifik, gunakan `/public/location`**
      operationId: getPublicWeather
      parameters:
        - name: province
          in: query
          description: Filter berdasarkan provinsi (gunakan underscore untuk spasi)
          schema:
            type: string
          example: jawa_tengah
        - name: kabupaten
          in: query
          description: Filter berdasarkan kabupaten/kota (gunakan underscore untuk spasi)
          schema:
            type: string
          example: banyumas
        - name: kecamatan
          in: query
          description: Filter berdasarkan kecamatan (gunakan underscore untuk spasi)
          schema:
            type: string
          example: purwokerto_selatan
        - name: format
          in: query
          description: |
            Format output:
            - `geojson` (default) - GeoJSON FeatureCollection dengan metadata
            - `json` - JSON array dengan wrapper
          schema:
            type: string
            enum: [geojson, json]
            default: geojson
      responses:
        '200':
          description: Data cuaca berhasil diambil
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/PublicWeatherJsonResponse'
                  - $ref: '#/components/schemas/PublicWeatherGeoJsonResponse'
            application/geo+json:
              schema:
                $ref: '#/components/schemas/PublicWeatherGeoJsonResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Error message here
        details:
          type: string
          description: Additional error details (optional)
        examples:
          type: object
          description: Usage examples (for 400 errors)
        supported:
          type: array
          items:
            type: string
          description: List of supported values (for 400 errors)
...
    NowcastingResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        source:
          type: string
          example: signature.bmkg.go.id
          description: Sumber data (signature.bmkg.go.id atau www.bmkg.go.id)
        code:
          type: string
          example: CJH
          description: Kode stasiun (untuk type=signature)
        province:
          type: string
          example: Jawa Tengah
          description: Nama provinsi (untuk type=xml/databmkg)
        data:
          type: object
          description: Data nowcasting dari BMKG
    PublicWeatherJsonResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        count:
          type: integer
          example: 150
          description: Jumlah data yang dikembalikan
        filters:
          type: object
          properties:
            province:
              type: string
              nullable: true
            kabupaten:
              type: string
              nullable: true
            kecamatan:
              type: string
              nullable: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/PublicWeatherFeature'
    PublicWeatherGeoJsonResponse:
      type: object
      properties:
        type:
          type: string
          enum: [FeatureCollection]
        features:
          type: array
          items:
            $ref: '#/components/schemas/PublicWeatherFeature'
        metadata:
          type: object
          properties:
            success:
              type: boolean
              example: true
            count:
              type: integer
              example: 150
            filters:
              type: object
              properties:
                province:
                  type: string
                  nullable: true
                kabupaten:
                  type: string
                  nullable: true
                kecamatan:
                  type: string
                  nullable: true
            generated:
              type: string
              format: date-time
              example: '2025-12-11T08:45:30.123Z'
    PublicWeatherFeature:
      type: object
      properties:
        type:
          type: string
          enum: [Feature]
        geometry:
          type: object
          properties:
            type:
              type: string
              enum: [Point]
            coordinates:
              type: array
              items:
                type: number
              minItems: 2
              maxItems: 2
        properties:
          type: object
          properties:
            id:
              type: string
            province:
              type: string
            kabupaten:
              type: string
            kecamatan:
              type: string
            timestamp:
              type: string
            type:
              type: string
            weather:
              type: object
              nullable: true
              description: Data cuaca dari BMKG
              properties:
                humidity:
                  type: string
                  description: Kelembapan udara (%)
                temperature:
                  type: string
                  description: Suhu udara (°C)
                weatherCode:
                  type: string
                  description: Kode cuaca (0-97), lihat kode.md untuk referensi
                windDirection:
                  type: string
                  description: Arah angin dari (N, NE, E, SE, S, SW, W, NW, dll)
                windSpeed:
                  type: string
                  description: Kecepatan angin (km/jam)
    LocationWeatherResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        source:
          type: string
          example: signature.bmkg.go.id
        type:
          type: string
          enum: [by_code, by_coordinates]
          example: by_code
        code:
          type: string
          description: Kode ADM4 (jika query by code)
          example: '33.01.22.1003'
        coordinates:
          type: object
          description: Koordinat (jika query by coordinates)
          properties:
            lat:
              type: number
              example: -7.656747
            lon:
              type: number
              example: 109.115523
        data:
          type: object
          description: Data cuaca dari BMKG
          properties:
            status:
              type: integer
              example: 200
            data:
              type: object
              properties:
                lokasi:
                  $ref: '#/components/schemas/BMKGLocationInfo'
                cuaca:
                  $ref: '#/components/schemas/BMKGCurrentWeather'
                prakiraan:
                  type: array
                  items:
                    $ref: '#/components/schemas/BMKGForecastWeather'
    BMKGLocationInfo:
      type: object
      description: Informasi lokasi dari BMKG
      properties:
        lon:
          type: number
          description: Longitude
          example: 109.015875
        lat:
          type: number
          description: Latitude
          example: -7.709959
        adm1:
          type: string
          description: Kode provinsi
          example: '33'
        adm2:
          type: string
          description: Kode kabupaten/kota
          example: '33.01'
        adm3:
          type: string
          description: Kode kecamatan
          example: '33.01.22'
        adm4:
          type: string
          description: Kode kelurahan/desa
          example: '33.01.22.1003'
        provinsi:
          type: string
          example: Jawa Tengah
        kotkab:
          type: string
          example: Cilacap
        kecamatan:
          type: string
          example: Cilacap tengah
        desa:
          type: string
          example: Sidanegara
    BMKGCurrentWeather:
      type: object
      description: Data cuaca saat ini
      properties:
        weather:
          type: integer
          description: Kode cuaca (0-104), lihat kode.md
          example: 3
        weather_desc:
          type: string
          description: Deskripsi cuaca (ID)
          example: Berawan
        weather_desc_en:
          type: string
          description: Deskripsi cuaca (EN)
          example: Mostly Cloudy
        image:
          type: string
          description: URL ikon cuaca dari BMKG
          example: https://api-apps.bmkg.go.id/storage/icon/cuaca/berawan-pm.svg
        datetime:
          type: string
          example: '2025-12-11T14:40:00.000Z'
        local_datetime:
          type: string
          example: '2025-12-11 21:40:00'
        t:
          type: number
          description: Suhu udara (°C)
          example: 26
        tcc:
          type: number
          description: Tutupan awan (%)
          example: 62
        wd_deg:
          type: number
          description: Arah angin dalam derajat
          example: 239
        wd:
          type: string
          description: Arah angin dari (N, NE, E, SE, S, SW, W, NW)
          example: SW
        wd_to:
          type: string
          description: Arah angin menuju
          example: NE
        ws:
          type: number
          description: Kecepatan angin (km/jam)
          example: 3.5
        hu:
          type: number
          description: Kelembapan udara (%)
          example: 89
        vs:
          type: number
          description: Jarak pandang (meter)
          example: 9467
        vs_text:
          type: string
          description: Jarak pandang (text)
          example: '< 10 km'
        source:
          type: string
          example: coord-pwx
    BMKGForecastWeather:
      type: object
      description: Data prakiraan cuaca
      allOf:
        - $ref: '#/components/schemas/BMKGCurrentWeather'
        - type: object
          properties:
            tp:
              type: number
              description: Total presipitasi (mm)
              example: 0.2
            time_index:
              type: string
              example: '2-3'
            analysis_date:
              type: string
              example: '2025-12-11T12:00:00'
            utc_datetime:
              type: string
              example: '2025-12-11 15:00:00'